+++ include/zephyr/drivers/wifi/nrf_wifi/bus/qspi_if.h
