+++ modules/nrf_wifi/bus/spi_nor.h
