+++ modules/nrf_wifi/os/timer.c
